flight: refactor event system into systemmod
Remove pipxtrememod, factored in term of system mod now.  Ensure that
periodic things like byte count calculations are done properly irrespective
of arming status.
