+++ flight/targets/aq32/fw/pios_board.c
@@ -245,7 +245,6 @@ void PIOS_Board_Init(void) {
     TaskMonitorInitialize();
 
     /* Initialize UAVObject libraries */
-    EventDispatcherInitialize();
     UAVObjInitialize();
 
     /* Initialize the alarms library */
