+++ flight/targets/coptercontrol/fw/Makefile
@@ -160,7 +160,6 @@ SRC += pios_board.c
 SRC += $(FLIGHTLIB)/alarms.c
 SRC += $(OPUAVTALK)/uavtalk.c
 SRC += $(OPUAVOBJ)/uavobjectmanager.c
-SRC += $(OPUAVOBJ)/eventdispatcher.c
 SRC += $(FLIGHTLIB)/msplib.c
 
 else
