+++ flight/targets/coptercontrol/fw/pios_board.c
@@ -144,7 +144,6 @@ void PIOS_Board_Init(void) {
 	TaskMonitorInitialize();
 
 	/* Initialize UAVObject libraries */
-	EventDispatcherInitialize();
 	UAVObjInitialize();
 
 	/* Initialize the alarms library. Reads RCC reset flags */
