+++ flight/targets/naze32/fw/pios_board.c
@@ -306,7 +306,6 @@ void PIOS_Board_Init(void) {
 #endif	/* PIOS_INCLUDE_FLASH */
 
 	/* Initialize UAVObject libraries */
-	EventDispatcherInitialize();
 	UAVObjInitialize();
 
 	/* Initialize the hardware UAVOs */
