+++ flight/targets/sparky/fw/pios_board.c
@@ -273,7 +273,6 @@ void PIOS_Board_Init(void)
 	TaskMonitorInitialize();
 
 	/* Initialize UAVObject libraries */
-	EventDispatcherInitialize();
 	UAVObjInitialize();
 
 	/* Initialize the alarms library. Reads RCC reset flags */
