+++ flight/targets/sparky2/fw/pios_board.c
@@ -444,7 +444,6 @@ void PIOS_Board_Init(void) {
 #endif	/* PIOS_INCLUDE_FLASH */
 
 	/* Initialize UAVObject libraries */
-	EventDispatcherInitialize();
 	UAVObjInitialize();
 
 	/* Initialize the hardware UAVOs */
