+++ flight/targets/sparkybgc/fw/pios_board.c
@@ -294,7 +294,6 @@ void PIOS_Board_Init(void) {
 	TaskMonitorInitialize();
 
 	/* Initialize UAVObject libraries */
-	EventDispatcherInitialize();
 	UAVObjInitialize();
 
 	/* Initialize the alarms library. Reads RCC reset flags */
